front end page split
